plugin: refactor loader
