+++ plugins/loader/loader_windows.go
@@ -1,4 +1,4 @@
-package plugins
+package loader
 
 import (
 	"fmt"
