+++ plugins/support/files.go
@@ -1,0 +1,12 @@
+package support
+
+import "runtime"
+
+// PluginExtension returns plugin extension format
+func PluginExtension() string {
+	if runtime.GOOS == "windows" {
+		return "dll"
+	}
+
+	return "so"
+}
